requirements.txt
google-generativeai
python-dotenv
streamlit
wikipedia
requests
beautifulsoup4
